shared styles between edit profile and edit hosting preferences page
